remove: Delete StoredPostBody component and integrate its functionality directly into StoredPostScreen
